Service Implementation done

+++ src/main/java/Sports_Manager/TeamManager/repos/MatchRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MatchRepo extends JpaRepository<Match,Integer> {
+public interface MatchRepo extends JpaRepository<Match,Long> {
     @Query(value = "SELECT m FROM Match m WHERE m.idmatches = ?1")
     public Match getByID(Long id);
 }
